requirements.txt
paramiko==2.3.1
Pillow==4.3.0
Wand==0.4.4
pygame==1.9.3
